Add char type, mutable strings, char stdlib; fix foreach counter
Language:
- New `char` primitive: 1-byte ASCII, distinct from u8, with `'x'` literals
  (escapes, non-ASCII rejected at lex) and char<->u8/u16 casts.
- A string is now semantically an array of `char`: `str[i]` and `for c in s`
  yield `char` (use `as u8` for the raw byte).
- Mutable fixed-capacity string buffers `str<N>`: own `1+N` bytes of RAM,
  initialized from a string literal, edited in place with `s[i] = c`. Resolves
  to the same runtime representation as `str`, so all str reads work unchanged.
  Writing through a plain `str` (ROM literal) is a compile error.

Stdlib:
- std/char.wr: ASCII ctype helpers (is_digit/is_alpha/is_upper/is_lower/
  is_alnum/is_whitespace, to_upper/to_lower, digit_value), all inline.
- std/math.wr: abs(i8) and abs16(i16).

Fix:
- ForEach over a string/array kept its loop counter only in X, which the inline
  u8 multiply (X as its shift counter) and nested loops clobbered, corrupting
  iteration and hanging the loop. The counter now lives in a hidden zero-page
  slot, reloaded into X at each loop head and advanced with INC, so the body
  may use X freely. Also fixes latently-broken nested foreach.

Tests: +regression coverage for all of the above (multiply/nested/indexed
foreach, buffer edits, char semantics, stdlib). Docs and examples updated.

+++ tests/e2e/complex_features.rs
@@ -251,8 +251,8 @@ fn string_len_and_index() {
         fn main() {
             let s: str = "hello";
             LEN = s.len as u8;
-            C0 = s[0];
-            C4 = s[4];
+            C0 = s[0] as u8;
+            C4 = s[4] as u8;
             loop {}
         }
     "#);
